requirements.txt
requests
python-dotenv
rasterio
scikit-learn
scipy

